requirements.txt
streamlit
pandas
plotly
scikit-learn
pytesseract
Pillow
fpdf2